+++ musicalgestures/_history.py
@@ -91,9 +91,12 @@ def history(self, filename='', history_length=10):
         # mg_progressbar(ii, length+1, 'Rendering history video:', 'Complete')
 
     out.release()
-    source_audio = extract_wav(self.of + self.fex)
+    
     destination_video = self.of + '_history' + self.fex
-    embed_audio_in_video(source_audio, destination_video)
-    os.remove(source_audio)
+
+    if self.has_audio:
+        source_audio = extract_wav(self.of + self.fex)
+        embed_audio_in_video(source_audio, destination_video)
+        os.remove(source_audio)
 
     return musicalgestures.MgObject(destination_video, color=self.color, returned_by_process=True)
